+++ integration_tests/suite/base/test_group_call_permission.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from hamcrest import assert_that, contains, empty, has_entries
@@ -19,6 +19,9 @@ def test_associate_errors(group, call_permission):
     yield s.check_resource_not_found, fake_group, 'Group'
     yield s.check_resource_not_found, fake_call_permission, 'CallPermission'
 
+    fake_call_permission = confd.groups(group['uuid']).callpermissions(FAKE_ID).put
+    yield s.check_resource_not_found, fake_call_permission, 'CallPermission'
+
 
 @fixtures.group()
 @fixtures.call_permission()
@@ -29,11 +32,17 @@ def test_dissociate_errors(group, call_permission):
     yield s.check_resource_not_found, fake_group, 'Group'
     yield s.check_resource_not_found, fake_call_permission, 'CallPermission'
 
+    fake_call_permission = confd.groups(group['uuid']).callpermissions(FAKE_ID).delete
+    yield s.check_resource_not_found, fake_call_permission, 'CallPermission'
+
 
 @fixtures.group()
 @fixtures.call_permission()
 def test_associate(group, call_permission):
     response = confd.groups(group['id']).callpermissions(call_permission['id']).put()
+    response.assert_updated()
+
+    response = confd.groups(group['uuid']).callpermissions(call_permission['id']).put()
     response.assert_updated()
 
 
@@ -42,8 +51,8 @@ def test_associate(group, call_permission):
 @fixtures.call_permission()
 @fixtures.call_permission()
 def test_associate_multiple_call_permissions_to_group(group, perm1, perm2, perm3):
-    confd.groups(group['id']).callpermissions(perm1['id']).put().assert_updated()
-    confd.groups(group['id']).callpermissions(perm2['id']).put().assert_updated()
+    confd.groups(group['uuid']).callpermissions(perm1['id']).put().assert_updated()
+    confd.groups(group['uuid']).callpermissions(perm2['id']).put().assert_updated()
     confd.groups(group['id']).callpermissions(perm3['id']).put().assert_updated()
 
 
@@ -54,10 +63,10 @@ def test_associate_multiple_call_permissions_to_group(group, perm1, perm2, perm3
 def test_associate_multiple_groups_to_call_permission(
     group1, group2, group3, call_permission
 ):
-    confd.groups(group1['id']).callpermissions(
+    confd.groups(group1['uuid']).callpermissions(
         call_permission['id']
     ).put().assert_updated()
-    confd.groups(group2['id']).callpermissions(
+    confd.groups(group2['uuid']).callpermissions(
         call_permission['id']
     ).put().assert_updated()
     confd.groups(group3['id']).callpermissions(
@@ -76,6 +85,12 @@ def test_associate_when_group_already_associated_to_same_call_permission(
         )
         response.assert_updated()
 
+    with a.group_call_permission(group, call_permission):
+        response = (
+            confd.groups(group['uuid']).callpermissions(call_permission['id']).put()
+        )
+        response.assert_updated()
+
 
 @fixtures.group(wazo_tenant=MAIN_TENANT)
 @fixtures.group(wazo_tenant=SUB_TENANT)
@@ -83,21 +98,21 @@ def test_associate_when_group_already_associated_to_same_call_permission(
 @fixtures.call_permission(wazo_tenant=SUB_TENANT)
 def test_associate_multi_tenant(main_group, sub_group, main_perm, sub_perm):
     response = (
-        confd.groups(main_group['id'])
+        confd.groups(main_group['uuid'])
         .callpermissions(sub_perm['id'])
         .put(wazo_tenant=SUB_TENANT)
     )
     response.assert_match(404, e.not_found('Group'))
 
     response = (
-        confd.groups(sub_group['id'])
+        confd.groups(sub_group['uuid'])
         .callpermissions(main_perm['id'])
         .put(wazo_tenant=SUB_TENANT)
     )
     response.assert_match(404, e.not_found('CallPermission'))
 
     response = (
-        confd.groups(main_group['id'])
+        confd.groups(main_group['uuid'])
         .callpermissions(sub_perm['id'])
         .put(wazo_tenant=MAIN_TENANT)
     )
@@ -107,6 +122,12 @@ def test_associate_multi_tenant(main_group, sub_group, main_perm, sub_perm):
 @fixtures.group()
 @fixtures.call_permission()
 def test_dissociate(group, call_permission):
+    with a.group_call_permission(group, call_permission, check=False):
+        response = (
+            confd.groups(group['uuid']).callpermissions(call_permission['id']).delete()
+        )
+        response.assert_deleted()
+
     with a.group_call_permission(group, call_permission, check=False):
         response = (
             confd.groups(group['id']).callpermissions(call_permission['id']).delete()
@@ -120,14 +141,14 @@ def test_dissociate(group, call_permission):
 @fixtures.call_permission(wazo_tenant=SUB_TENANT)
 def test_dissociate_multi_tenant(main_group, sub_group, main_perm, sub_perm):
     response = (
-        confd.groups(main_group['id'])
+        confd.groups(main_group['uuid'])
         .callpermissions(sub_perm['id'])
         .delete(wazo_tenant=SUB_TENANT)
     )
     response.assert_match(404, e.not_found('Group'))
 
     response = (
-        confd.groups(sub_group['id'])
+        confd.groups(sub_group['uuid'])
         .callpermissions(main_perm['id'])
         .delete(wazo_tenant=SUB_TENANT)
     )
@@ -138,7 +159,7 @@ def test_dissociate_multi_tenant(main_group, sub_group, main_perm, sub_perm):
 @fixtures.call_permission()
 def test_get_call_permissions_relation(group, call_permission):
     with a.group_call_permission(group, call_permission):
-        response = confd.groups(group['id']).get()
+        response = confd.groups(group['uuid']).get()
         assert_that(
             response.item['call_permissions'],
             contains(
@@ -154,7 +175,9 @@ def test_get_groups_relation(group, call_permission):
         response = confd.callpermissions(call_permission['id']).get()
         assert_that(
             response.item['groups'],
-            contains(has_entries(id=group['id'], name=group['name'])),
+            contains(
+                has_entries(uuid=group['uuid'], id=group['id'], name=group['name'])
+            ),
         )
 
 
@@ -162,7 +185,7 @@ def test_get_groups_relation(group, call_permission):
 @fixtures.call_permission()
 def test_delete_group_when_group_and_call_permission_associated(group, call_permission):
     with a.group_call_permission(group, call_permission, check=False):
-        confd.groups(group['id']).delete().assert_deleted()
+        confd.groups(group['uuid']).delete().assert_deleted()
         response = confd.callpermissions(call_permission['id']).get()
         assert_that(response.item['groups'], empty())
 
@@ -180,14 +203,25 @@ def test_delete_call_permission_when_group_and_call_permission_associated(
 
 @fixtures.group()
 @fixtures.call_permission()
+def test_delete_call_permission_when_group_and_call_permission_associated_by_uuid(
+    group, call_permission
+):
+    with a.group_call_permission(group, call_permission, check=False):
+        confd.callpermissions(call_permission['id']).delete().assert_deleted()
+        response = confd.groups(group['uuid']).get()
+        assert_that(response.item['call_permissions'], empty())
+
+
+@fixtures.group()
+@fixtures.call_permission()
 def test_bus_events(group, call_permission):
     yield (
         s.check_bus_event,
-        'config.groups.{}.callpermissions.updated'.format(group['id']),
-        confd.groups(group['id']).callpermissions(call_permission['id']).put,
+        'config.groups.*.callpermissions.updated',
+        confd.groups(group['uuid']).callpermissions(call_permission['id']).put,
     )
     yield (
         s.check_bus_event,
-        'config.groups.{}.callpermissions.deleted'.format(group['id']),
-        confd.groups(group['id']).callpermissions(call_permission['id']).delete,
+        'config.groups.*.callpermissions.deleted',
+        confd.groups(group['uuid']).callpermissions(call_permission['id']).delete,
     )