+++ wazo_confd/plugins/call_permission/schema.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2018-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from marshmallow import fields
@@ -26,7 +26,7 @@ class CallPermissionSchema(BaseSchema):
         'OutcallSchema', only=['id', 'name', 'links'], many=True, dump_only=True
     )
     groups = fields.Nested(
-        'GroupSchema', only=['id', 'name', 'links'], many=True, dump_only=True
+        'GroupSchema', only=['uuid', 'id', 'name', 'links'], many=True, dump_only=True
     )
     users = fields.Nested(
         'UserSchema',