+++ wazo_confd/plugins/group/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .resource import GroupItem, GroupList
@@ -14,7 +14,8 @@ class Plugin:
 
         api.add_resource(
             GroupItem,
-            '/groups/<int:id>',
+            '/groups/<uuid:uuid>',
+            '/groups/<int:uuid>',
             endpoint='groups',
             resource_class_args=(service,),
         )