+++ wazo_confd/plugins/group_call_permission/notifier.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo_bus.resources.group_call_permission.event import (
@@ -14,11 +14,19 @@ class GroupCallPermissionNotifier:
         self.bus = bus
 
     def associated(self, group, call_permission):
-        event = GroupCallPermissionAssociatedEvent(group.id, call_permission.id)
+        event = GroupCallPermissionAssociatedEvent(
+            group_id=group.id,
+            group_uuid=str(group.uuid),
+            call_permission_id=call_permission.id,
+        )
         self.bus.send_bus_event(event)
 
     def dissociated(self, group, call_permission):
-        event = GroupCallPermissionDissociatedEvent(group.id, call_permission.id)
+        event = GroupCallPermissionDissociatedEvent(
+            group_id=group.id,
+            group_uuid=str(group.uuid),
+            call_permission_id=call_permission.id,
+        )
         self.bus.send_bus_event(event)