+++ wazo_confd/plugins/group_fallback/resource.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import request
@@ -18,14 +18,14 @@ class GroupFallbackList(ConfdResource):
         self.service = service
         self.group_dao = group_dao
 
-    @required_acl('confd.groups.{group_id}.fallbacks.read')
-    def get(self, group_id):
-        group = self.group_dao.get(group_id)
+    @required_acl('confd.groups.{group_uuid}.fallbacks.read')
+    def get(self, group_uuid):
+        group = self.group_dao.get(group_uuid)
         return self.schema().dump(group.fallbacks)
 
-    @required_acl('confd.groups.{group_id}.fallbacks.update')
-    def put(self, group_id):
-        group = self.group_dao.get(group_id)
+    @required_acl('confd.groups.{group_uuid}.fallbacks.update')
+    def put(self, group_uuid):
+        group = self.group_dao.get(group_uuid)
         fallbacks = self.schema().load(request.get_json())
         self.service.edit(group, fallbacks)
         return '', 204