+++ wazo_confd/plugins/group_schedule/notifier.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo_bus.resources.group_schedule.event import (
@@ -14,11 +14,19 @@ class GroupScheduleNotifier:
         self.bus = bus
 
     def associated(self, group, schedule):
-        event = GroupScheduleAssociatedEvent(group.id, schedule.id)
+        event = GroupScheduleAssociatedEvent(
+            group_id=group.id,
+            group_uuid=str(group.uuid),
+            schedule_id=schedule.id,
+        )
         self.bus.send_bus_event(event)
 
     def dissociated(self, group, schedule):
-        event = GroupScheduleDissociatedEvent(group.id, schedule.id)
+        event = GroupScheduleDissociatedEvent(
+            group_id=group.id,
+            group_uuid=str(group.uuid),
+            schedule_id=schedule.id,
+        )
         self.bus.send_bus_event(event)