+++ wazo_confd/plugins/group_schedule/validator.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo_dao.helpers import errors
@@ -23,7 +23,7 @@ class GroupScheduleAssociationValidator(ValidatorAssociation):
             raise errors.resource_associated(
                 'Group',
                 'Schedule',
-                group_id=group.id,
+                group_uuid=group.uuid,
                 schedule_id=group.schedules[0].id,
             )